requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.8
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.7.0
requests>=2.32.0
feedparser>=6.0.0
apscheduler>=3.10.0
newsapi-python>=0.2.7
ta>=0.11.0
httpx>=0.27.0